# docs/swagger.yaml
basePath: /v1
definitions:
...
  main.VenueDetailResponse:
    properties:
      address:
        type: string
      amenities:
        items:
          type: string
        type: array
      average_rating:
        type: number
      completed_games:
        type: integer
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      image_urls:
        items:
          type: string
        type: array
      location:
        description: '[latitude, longitude]'
        items:
          type: number
        type: array
      name:
        type: string
      open_time:
        type: string
      owner_id:
        type: integer
      phone_number:
        type: string
      sport:
        type: string
      total_reviews:
        type: integer
      upcoming_games:
        type: integer
      updated_at:
        type: string
    type: object
  main.VenueListResponse:
    properties:
      address:
        type: string
      average_rating:
        type: number
      id:
        type: integer
      image_urls:
        items:
          type: string
        type: array
      is_favorite:
        type: boolean
      location:
        description: '[longitude, latitude]'
        items:
          type: number
        type: array
      name:
        type: string
      open_time:
        type: string
      phone_number:
        type: string
      sport:
        type: string
      total_reviews:
        type: integer
    type: object
...
  main.isOwnerResponse:
    properties:
      is_owner:
        type: boolean
      venue_ids:
        items:
          type: integer
        type: array
    type: object
...
  venues.Venue:
    properties:
      address:
        type: string
      amenities:
        description: Array of strings
        items:
          type: string
        type: array
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      image_urls:
        description: Array of image URLs
        items:
          type: string
        type: array
      location:
        description: PostGIS point (longitude, latitude)
        items:
          type: number
        type: array
      name:
        type: string
      open_time:
        type: string
      owner_id:
        type: integer
      phone_number:
        type: string
      sport:
        type: string
      updated_at:
        type: string
    type: object
  venues.VenueInfo:
    properties:
      address:
        type: string
      amenities:
        description: Array of strings
        items:
          type: string
        type: array
      description:
        type: string
      id:
        type: integer
      location:
        description: PostGIS point (longitude, latitude)
        items:
          type: number
        type: array
      name:
        type: string
      open_time:
        type: string
      phone_number:
        type: string
      status:
        type: string
    type: object
info:
  contact:
    email: Ongchen10sherpa@gmail.com
    name: fullstacksherpa
    url: https://www.fullstacksherpa.tech/
  description: API for Khel, a complete sport application.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Khel API
paths:
...
  /venue/{id}:
    get:
      description: Retrieve detailed information for a venue including aggregated
        review and game statistics.
      parameters:
      - description: Venue ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successful response with venue details
          schema:
            $ref: '#/definitions/main.VenueDetailResponse'
        "400":
          description: 'Bad Request: Invalid venue id'
          schema: {}
        "404":
          description: 'Not Found: Venue not found'
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Get venue details
      tags:
      - Venue
  /venues:
    post:
      consumes:
      - multipart/form-data
      description: Register a new venue with details such as name, address, location,
        and amenities.
      parameters:
      - description: Venue details (JSON string)
        in: formData
        name: venue
        required: true
        type: string
      - collectionFormat: csv
        description: Venue images (up to 7 files)
        in: formData
        items:
          type: file
        name: images
        type: array
      produces:
      - application/json
      responses:
        "201":
          description: Venue created successfully
          schema:
            $ref: '#/definitions/venues.Venue'
        "400":
          description: Invalid request payload
          schema: {}
        "401":
          description: Unauthorized
          schema: {}
        "500":
          description: Internal server error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Register a venue in our system
      tags:
      - Venue-Owner
  /venues/{venueID}:
    delete:
      description: Deletes a venue by ID and removes all associated images from Cloudinary.
      parameters:
      - description: Venue ID
        in: path
        name: venueID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Venue deleted successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid venue ID
          schema: {}
        "401":
          description: Unauthorized
          schema: {}
        "404":
          description: Venue not found
          schema: {}
        "500":
          description: Internal server error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Delete a venue from the system
      tags:
      - Venue-Owner
    patch:
      consumes:
      - application/json
      description: Allows venue owners to update partial information about their venue.
      parameters:
      - description: Venue ID
        in: path
        name: venueID
        required: true
        type: integer
      - description: Venue update payload
        in: body
        name: updateData
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Venue updated successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 'Bad Request: Invalid input'
          schema: {}
        "500":
          description: 'Internal Server Error: Could not update venue'
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Update venue information
      tags:
      - Venue-Owner
...
  /venues/{venueID}/photos:
    delete:
      consumes:
      - application/json
      description: Deletes a specific venue photo from Cloudinary and removes it from
        the database.
      parameters:
      - description: Venue ID
        in: path
        name: venueID
        required: true
        type: integer
      - description: Photo URL to delete
        in: query
        name: photo_url
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Photo deleted successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 'Bad Request: Missing venue ID or photo URL'
          schema: {}
        "500":
          description: 'Internal Server Error: Could not delete photo'
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Delete a venue photo
      tags:
      - Venue-Owner
    get:
      description: Returns a list of all image URLs associated with the specified
        venue.
      parameters:
      - description: Venue ID
        in: path
        name: venueID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of image URLs
          schema:
            items:
              type: string
            type: array
        "400":
          description: Invalid venue ID
          schema: {}
        "401":
          description: Unauthorized
          schema: {}
        "404":
          description: Venue not found
          schema: {}
        "500":
          description: Internal server error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Retrieve all venue photo URLs
      tags:
      - Venue-Owner
    post:
      consumes:
      - multipart/form-data
      description: Uploads a new venue photo to Cloudinary and adds the new photo
        URL to the venue record.
      parameters:
      - description: Venue ID
        in: path
        name: venueID
        required: true
        type: integer
      - description: Photo file to upload
        in: formData
        name: photo
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: 'Photo uploaded successfully, returns {\"photo_url\": \"<newPhotoURL>\"}'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 'Bad Request: Invalid input or missing file'
          schema: {}
        "500":
          description: 'Internal Server Error: Could not process the upload'
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Upload a new photo for a venue
      tags:
      - Venue-Owner
...
  /venues/{venueID}/venue-info:
    get:
      description: Retrieves detailed information (name, address, description, status,
        etc.) about a venue by its ID.
      parameters:
      - description: Venue ID
        in: path
        name: venueID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Detailed venue information
          schema:
            $ref: '#/definitions/venues.VenueInfo'
        "400":
          description: Invalid venue ID
          schema: {}
        "401":
          description: Unauthorized
          schema: {}
        "404":
          description: Venue not found
          schema: {}
        "500":
          description: Internal server error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Get venue information
      tags:
      - Venue-Owner
...
  /venues/is-venue-owner:
    get:
      description: Determines whether the authenticated user owns at least one venue
      produces:
      - application/json
      responses:
        "200":
          description: Ownership check result
          schema:
            $ref: '#/definitions/main.isOwnerResponse'
        "401":
          description: Unauthorized
          schema: {}
        "500":
          description: Internal server error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Check if user is a venue owner
      tags:
      - Venue-Owner
  /venues/list-venues:
    get:
      consumes:
      - application/json
      description: Get paginated list of venues with filters
      parameters:
      - description: Filter by sport type
        in: query
        name: sport
        type: string
      - description: Latitude for location filter
        in: query
        name: lat
        type: number
      - description: Longitude for location filter
        in: query
        name: lng
        type: number
      - description: Distance in meters from location
        in: query
        name: distance
        type: number
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 7
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.VenueListResponse'
            type: array
      security:
      - ApiKeyAuth: []
      summary: List venues
      tags:
      - Venue
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"